Fix a couple of compiler warnings related to type conversion from chars to ints for isspace(), isdigit() etc.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@549225 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_ajp12_worker.c
@@ -542,7 +542,7 @@ static int ajpv12_handle_response(ajp12_endpoint_t * p,
         }
 
         name = line;
-        while (isspace(*name) && *name) {
+        while (isspace((int)(*name)) && *name) {
             name++;             /* Skip leading white chars */
         }
         if (!*name) {           /* Empty header name */
@@ -557,7 +557,7 @@ static int ajpv12_handle_response(ajp12_endpoint_t * p,
         }
         *value = '\0';
         value++;
-        while (isspace(*value) && *value) {
+        while (isspace((int)(*value)) && *value) {
             value++;            /* Skip leading white chars */
         }
         if (!*value) {          /* Empty header value */
